Fixed Multiple-Unit Placement
The amount of units placed will be equal to the amount in "All Entities Prefab". Currently, units are selected by random from that array, so the number of times a unit is in the array directly correlates to their pick rate. I tried to implement a non-random instantiation option, but for some reason it causes Team2 to instantly die.
The health bar size of the archer is also too big, but I think that's because actual dimensions of the sprite image are smaller than the knight. I'm not sure if we can adjust the size per sprite, so you may need to find a different sprite or adjust its dimensions manually.
We might be good with just fixing what I mentioned, but if you really want to add more, I would prioritize basic user interaction / unit placement. However, we should probably try to focus on the project report instead.